chore(tile.py): move `tile.py` -> `.../surfaces/game_map/tiles/`
